refactor : projection dto 로 디렉토리 변경
