+++ src/main/java/club/inq/team1/dto/projection/FollowerUserProjectionDTO.java
@@ -1,4 +1,4 @@
-package club.inq.team1.repository;
+package club.inq.team1.dto.projection;
 
 import club.inq.team1.entity.User;
